4.1: renamed component
